Use decimal type for area to easy compare values

+++ sql.sql
@@ -6,6 +6,6 @@ CREATE TABLE district (
     district_id int NOT NULL PRIMARY KEY AUTO_INCREMENT,
     name varchar(50) NOT NULL,
     population int NOT NULL,
-    area float NOT NULL,
+    area decimal(5,2) NOT NULL,
     city varchar(50) NOT NULL
 ) ENGINE = InnoDB DEFAULT CHARSET = utf8 COLLATE utf8_polish_ci;
